fix(BudgetTest): fixed tests based on changes of the constructor in the Budget class.

+++ src/test/java/no/ntnu/idatt1002/demo/data/BudgetTest.java
@@ -2,6 +2,8 @@ package no.ntnu.idatt1002.demo.data;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class BudgetTest {
@@ -9,7 +11,8 @@ class BudgetTest {
   @Test
   void add() {
     Budget budget = new Budget();
-    budget.add(ExpenseType.TYPE6,123);
-    assertEquals(123,budget.getAmount(ExpenseType.TYPE6));
+
+    budget.add(ExpenseType.TYPE6, BigDecimal.valueOf(123));
+    assertEquals(BigDecimal.valueOf(123),budget.getAmount(ExpenseType.TYPE6));
   }
 }
